Safari-609.1.11.1/LayoutTests/platform/mac-wk2/plugins/npruntime/remove-property-from-javascript-expected.txt


Test that 'delete object.property' ends up calling NP_RemoveProperty

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testObject.lastRemovedProperty is 'testProperty'
PASS testObject.lastRemovedProperty is 242
